% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalStats.R
\name{additionalStats}
\alias{additionalStats}
\title{Generate miscellaneous statistics for species range shifts}
\usage{
additionalStats(
  result_dir,
  time_periods,
  scenarios,
  dispersal = FALSE,
  dispersaldata = NA,
  ncores = 1
)
}
\arguments{
\item{result_dir}{the directory where the ensembled and binary maps are placed in
addition to the "Results.csv" file. If \code{projectSuit} was used to make
these maps, this should be the same as the \code{output} argument in that function.}

\item{time_periods}{a vector of the years in which the projection will occur, with the
first element as the year the model will be trained on (usually the current data).If no
precise years are available (e.g., using data from the Last Glacial Maximum), order
time periods from current to least current and give character strings for the years (e.g., "LGM").}

\item{scenarios}{a vector of character strings detailing the different climate models
used in the forecasted/hindcasted species distribution models. In no projection is
needed, set to NA (defualt).}

\item{dispersal}{(logical \code{TRUE} or \code{FALSE}) Should these statistics be
calculated for the dispersal-constrained distribution maps? If dispersal rate
analysis are not needed, or if the \code{megaSDM::dispersalRate} function has yet
to be run, this should be set to \code{FALSE} (the default).}

\item{ncores}{the number of computer cores to parallelize the background point
generation on. Default is 1; Using one fewer core than the computer has is usually
optimal.}

\item{disperaldata}{A dataframe or the full file path to a .csv file with two columns:
\enumerate{
\item Species.
\item Dispersal Rate in km/yr.
See the function \code{megaSDM::dispersalRate} for more details.
}}
}
\value{
creates .pdf files of graphs showing changes in range size and distribution
across multiple time periods and scenarios:
\enumerate{
\item The overall modelled range size across all time periods and scenarios.
\item The percent change from the current range size.
\item Average range size for each year given multiple climate scenarios.
\item (If \code{dispersal = TRUE}) the difference in range size between dispersal constrained
and non-dispersal constrained species ranges.
}
}
\description{
This function generates graphs showing the changes in range size and position between
the data a species distribution model was trained on and future or past projections
of species ranges.
}
\details{
NOTE: this function is dependent on the outputs generated by the \code{projectSuit}
function, in particular the "Results.csv" file, which includes information on
changes in range size and position across the projected time periods and scenarios.
}
