% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTimeMaps.R
\name{createTimeMaps}
\alias{createTimeMaps}
\title{Create maps describing species range shifts across many time periods}
\usage{
createTimeMaps(
  result_dir,
  time_periods,
  scenarios,
  dispersal,
  dispersaldata,
  ncores
)
}
\arguments{
\item{result_dir}{the directory where the ensembled and binary maps are placed.
Each species should have its own sub-directory, and the forecasted/hindcasted
binary maps should be placed into directories like so: Species/Scenario/Time.
If \code{projectSuit} was used to make these maps, this is probably the same
as the \code{output} argument in that function.}

\item{time_periods}{a vector of the years in which the projection will occur. The first
element should be the original year (the year in which the model was generated). If no
precise years are available (e.g., using data from the Last Glacial Maximum), order data
from current to least current (furthest into the future/past) and give character strings
for the years (e.g., "LGM").}

\item{scenarios}{a vector of character strings detailing the different climate models
used in the forecasted/hindcasted species distribution models.}

\item{dispersal}{(logical \code{TRUE} or \code{FALSE}) Are the binary maps given
dispersal-constrained maps or non-dispersal-constrained maps? If dispersal rate
analysis are not needed, or if the \code{megaSDM::dispersalRate} function has yet
to be run, this should be set to \code{FALSE} (the default). NOTE: if \code{dispersal = TRUE},
then \code{time_periods} have to be numeric.}

\item{dispersaldata}{either a dataframe or the complete path name of a .csv file with two columns:
Column 1: species name (same as the name used for modelling)
Column 2: average dispersal rate of species in kilometers/year.}

\item{ncores}{the number of computer cores to parallelize the background point generation on.
Default is 1; Using one fewer core than the computer has is usually optimal.}
}
\value{
writes both rasters and pdfs of the maps showing the expansion and contraction
of each species range across all time-steps for each cliamte scenario.
}
\description{
Creates maps describing species range shifts across multiple time periods. These
maps detail the step-wise expansions and contractions of the species distribution
through those time-steps, allowing for the visualization of both unidirectional
range shifts (e.g., a range expansion across all time-steps) and more complex
dynamics (e.g., a range expansion followed by a range contraction).
}
