\name{megaSDM-package}
\alias{megaSDM-package}
\alias{megaSDM}
\docType{package}
\title{
\packageTitle{megaSDM}
}
\description{
\packageDescription{megaSDM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{megaSDM}
\packageIndices{megaSDM}
~~ An overview of how to use the package, including the most important functions ~~
}