% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OccurrenceCollection.R
\name{OccurrenceCollection}
\alias{OccurrenceCollection}
\title{Download and vet GBIF occurrence data}
\usage{
OccurrenceCollection(spplist, output, trainingarea = NA, studyarea = NA)
}
\arguments{
\item{spplist}{a vector of scientific names, using GBIF taxonomy. Names can be species or subspecies.}

\item{output}{A directory name where the downloaded species occurrences will be written to.}

\item{trainingarea}{Extent object, or vector of desired training extent in form \code{c(xmin, xmax, ymin, ymax)}.
Given in latlong coordinates. If set to "NA" (the default), all occurrence points will be
downloaded (up to 99,999 points), regardless of their location.}

\item{studyarea}{(optional) Extent object, or vector of the area that the SDM will be projected on in
form \code{c(xmin, xmax,ymin, ymax)}. Given in latlong coordinates. If provided, the number of occurrence
points found within this study region will be calculated.}
}
\value{
Writes .csv files of GBIF occurrences to a directory provided by \code{output}. If any species failed
(for example, the scientific name was not found in the search or no occurrences exist within a provided \code{trainingarea}),
another .csv file is written out in the same folder with the names of the species that failed. In addition,
a dataframe is returned by the function that contains the high taxonomy of each species and the nubmer of
occurrences found within the \code{trainingarea} and, optionally, the \code{studyarea}.
}
\description{
Takes a list of species and collects occurrence data from GBIF (Global Biodiversity Information Facility).
Acts as a wrapper for \code{rgbif::occ_search}; however, this function is much more efficient for a large
number of species. It also checks the taxonomy of the given species list against the GBIF taxonomy,
renaming or merging taxa if necessary. Furthermore, this function vets the occurrence data, removing
occurrence points that are of insufficient quality for species distribution modelling. Finally, \code{OccurrenceCollection()}
provides the number of occurrences found within given training and study areas. For a full list of
issues removed by this package, refer to <\emph{\strong{\strong{papername}}}__>. Further vetting may be done by hand.
}
