% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OccurrenceManagement.R
\name{OccurrenceManagement}
\alias{OccurrenceManagement}
\title{Manage and environmentally filter occurrence points}
\usage{
OccurrenceManagement(
  occlist,
  output,
  envextract = FALSE,
  envsample = FALSE,
  envdata,
  nbins = 25,
  PCA = "Y"
)
}
\arguments{
\item{occlist}{a list of .csv file names, each containing the occurrence
points of a given taxon. The files should be named the same as the taxon
of interest (e.g.,: ".../Canis_lupus.csv"). The occurrence points should
be given in latlong coordinates (-90 to 90, -180 to 180).}

\item{output}{the directory where the managed occurrence points will be placed.}

\item{envextract}{(logical (\code{TRUE} or \code{FALSE})) should the
environments at each occurrence point be extracted?}

\item{envsample}{(logical (\code{TRUE} or \code{FALSE})) should
environmental (Varela) subsampling be conducted on the occurrence points?}

\item{envdata}{a RasterStack or list of raster files
corresponding to the area the model will be trained on. All environmental
variables should be provided.}

\item{nbins}{(integer) the number of bins each climate combination
will be placed in for Varela subsampling of occurrence points. See
Varela et al. (2014) and Castellanos et al. (2018) for discussion on this
method. Default is 25 bins.}

\item{PCA}{(optional). The number of PC axes to use when binning climate
data for environmental subsampling. Can be a number from 1 to the number
of environmental variables. Default selects the number of axes that account
for 95\% of the environmental variation. If set to \code{NA}, PCA will not be
run (for use when there are categorical variables, for instance).}
}
\value{
a set of .csv files of occurrence points in the directory indicated by
the \code{output} argument with columns relating to the species name, the coordinates
(in the projection of the training area rasters), and the value of each
environmental layer at each point.
}
\description{
This function takes a set of occurrence points (whether downloaded from
GBIF using \code{OccurrenceCollection} or provided by the user),
standardizes the column headings for effective use in species distribution
modelling, and, if requested, extracts the values of each environmental
variable used in the modelling for each point and environmentally subsamples
the data using the method developed by Varela et al. (2014).
}
