% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableEnv.R
\name{VariableEnv}
\alias{VariableEnv}
\title{Use species-specific sets of environmental data for SDMs}
\usage{
VariableEnv(occlist, bglist, env_vars, occ_output, bg_output)
}
\arguments{
\item{occlist}{a list of .csv files, each containing the occurrence
points of a given taxon. The files should be named the same as the taxon
of interest (e.g.,: ".../Canis_lupus.csv").}

\item{bglist}{a list of .csv files corresponding to the background points
provided for each taxon. This list of files should be order in the same
way as the files given by \code{occlist}.}

\item{env_vars}{a vector containing the names of the environmental variables
(must exactly match column names) to keep for each species as a character
strings separated by commas. For example, \code{c("Bio1,Bio3", "Bio1,Bio12")}.
Like above, this vector should be in the same order as the lists of
species/background points.}

\item{occ_output}{the directory where output occurrence files will be placed}

\item{bg_output}{the directory where output background files will be placed}
}
\value{
Writes .csv files of species occurrences to the sub-directory \code{occ_output}
and .csv fiels of background points to the sub-directory \code{bg_output}.
}
\description{
Using different sets of environmental variables to model species
distributions can help to make more informative species distribution
models. This function allows for the modelling of each species based
upon a unique subset of the environmental variables.
}
