\name{megaSDM-package}
\alias{megaSDM-package}
\alias{megaSDM}
\docType{package}
\title{
\packageTitle{megaSDM}
}
\description{
\packageDescription{megaSDM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{megaSDM}
\packageIndices{megaSDM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{megaSDM}

Maintainer: \packageMaintainer{megaSDM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
